Temporarily move canvasRectangleToWindowRectangle to strategy
